feat: 모범답안 셋팅 및 부트 설정

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -5,28 +5,26 @@ import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.UserService;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.NoSuchElementException;
 import java.util.UUID;
 
 public class BasicUserService implements UserService {
-
     private final UserRepository userRepository;
 
-    public  BasicUserService(UserRepository userRepository) {
+    public BasicUserService(UserRepository userRepository) {
         this.userRepository = userRepository;
     }
 
     @Override
-    public User create(User user) {
-        if (user == null || user.getUserId() == null) {
-            throw new IllegalArgumentException("User 및 user ID는 null이 될 수 없습니다.");
-        }
-        return userRepository.create(user);
+    public User create(String username, String email, String password) {
+        User user = new User(username, email, password);
+        return userRepository.save(user);
     }
 
     @Override
-    public Optional<User> findById(UUID userId) {
-        return userRepository.findById(userId);
+    public User find(UUID userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
     }
 
     @Override
@@ -35,12 +33,18 @@ public class BasicUserService implements UserService {
     }
 
     @Override
-    public Optional<User> updateId(UUID userId, User updateUser) {
-        return userRepository.updateId(userId, updateUser);
+    public User update(UUID userId, String newUsername, String newEmail, String newPassword) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new NoSuchElementException("User with id " + userId + " not found"));
+        user.update(newUsername, newEmail, newPassword);
+        return userRepository.save(user);
     }
 
     @Override
-    public boolean deleteById(UUID userId) {
-        return userRepository.deleteById(userId);
+    public void delete(UUID userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new NoSuchElementException("User with id " + userId + " not found");
+        }
+        userRepository.deleteById(userId);
     }
 }
